Criando beck do index

+++ malune/src/main/java/com/example/malune/repository/ProdutoRepository.java
@@ -1,14 +1,15 @@
 package com.example.malune.repository;
 
-import com.example.malune.entity.Categoria;
-import com.example.malune.entity.Produto;
+import java.util.List;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
+import com.example.malune.entity.Categoria;
+import com.example.malune.entity.Produto;
 
 @Repository
-public interface ProdutoRepository extends JpaRepository<Produto, Long> {
+public interface ProdutoRepository extends JpaRepository<Produto, Integer> {
 
     List<Produto> findByCategoria(Categoria categoria);
